Dev (#80)
* remove the button in view staff

* edit staff profile

* delete  child- parser error, view staff-alert

* social worker alert

* parent-delete alert

* front end changes

---------

+++ src/components/staff/ViewStaff.jsx
@@ -20,9 +20,14 @@ const ViewStaff = () => {
   const [deleteRole] = useDeleteStaffProfileMutation();
 
   const handleDelete = async (roleId) => {
-    const resp = await deleteRole({ userIdToDelete: roleId });
-    if (isError) toast.error(resp.data.message);
-    if (isSuccess) refetch();
+    if (confirm('Are you sure?')) {
+      const resp = await deleteRole({ userIdToDelete: roleId });
+      if (isError) toast.error(resp.data.message);
+      if (isSuccess) refetch();
+    } else {
+      //action cancelled
+      console.log('Cancel is clicked.');
+    }
   };
 
   useEffect(() => {
